pyproject.toml
[project]
name = "shrtn"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "flask>=3.1.0",
    "gunicorn>=23.0.0",
    "python-dotenv>=1.0.1",
]
